Add a BindUntyped that seriously simplifies extending more
